:recycle: refactor: 파일명을 index로 변경
