+++ src/components/Tooltip/Tootip.stories.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Tooltip from './Tooltip';
+import Tooltip from '.';
 
 export default {
   title: 'CardAddPage/Tooltip',
