+++ src/components/common/Button/Button.stories.js
@@ -1,4 +1,4 @@
-import Button from './Button';
+import Button from '.';
 
 export default {
   title: 'Common/Button',
