+++ src/components/common/NoticeMessage/index.jsx
